+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.2
+- Fixed `RestrrOptions` and added `isWeb` attribute
+- Removed network check when no CookieJar is set (web)
+
 ## 0.3.1 
 - Added ability to customize & disable Cookie Jar